build: use pkg-config for X11
According to Daniel, AC_PATH_XTRA is sort-of deprecated.

This patch changes the configure arguments, --x-includes=dir and
--x-libraries=dir, which are now removed and use standard pkg-config.

It also replaces --with{,out}-x with --{dis,en}able-x11, which is the
same as other optionnal dependencies.

(this patch was done in an attempt to solve a build issue on maemo, it
turns out it didn't help)