--check: Updated manpage slightly and pulseaudio --help slightly.
It's good for new users - I didn't know about the convention for
daemon checks to only bump the exit code until, say, a year after I
first installed Linux. Of course, we could also put an entire guide to
using Linux in our manpage, or tell users how to check the exit code
with $?, or how to get to the terminal, or an example script, ... (you
can see how this would quickly get out of control). So just a little
bit more hint should be good for now. John? Does this work for you?

Sean

Signed-off-by: Lennart Poettering <lennart@poettering.net>