Create only the directory the current target should be created into.
With this change, instead of running multiple `mkdir -p` commands in all
the rules for all the generated definition files, only the call for the
current generated file is executed.

Not only it should shorten build time (especially for parallel make) but it
also fixes out-of-tree builds when new directories are added.

The $(dir $@) construct is GNU make-specific, but the rest of the
buildsystem is already GNU make-dependent so there should be no problem.