Allow to opt-out from building tests.
Since the tests are only useful either if you're hacking at pulseaudio as
a developer, or when running "make check", allow users to opt-out from
their build.

This for instance allows for Gentoo users not to build the tests when
installing the ebuild with tests disabled, and also allow for skipping over
eventually broken tests when trying to get the basic build going on a port.