Remove support for internal distributing and bundling of libltdl.
Standing to what the libtool documentation says, the LTDL_INIT macro and
the related configure options are only useful when intending to distribute
libltdl, and is superfluous if the system copy were always to be used.

Which makes it very easy to just drop the internal copy and use the system
library, just do it like any other library lacking pkg-config files to
identify its presence.

If this tries to build against an older libtool version it might fail at
link time, so for now this is not an user-proof solution. But it at least
should provide a working environment for packagers.