bluetooth: reinitialize the sample spec when switching profile
When doing init_profile(), the value u->sample_spec is modified to the
one which has been last configured. In case of HSP, it will be 8kHz.

Later on, when setting the profile to A2DP, it picks up the lower rate
available which match with current u->sample_spec.

In my case, it would be 16kHz. To circunvent the issue, I decided to
reinitialize the u->sample_spec to default value with user module
argument requested rate.