Use the _ONCE variation of AC_CHECKs where applicable.
When we're unconditionally looking for a function, or an header file,
without particular libraries set, or with found/not-found conditional code,
we can save ~1K lines in the final configure script by using the _ONCE
variant.

This makes sure that for each header of function, the check is done exactly
once and never more, reducing the amount of code that has to be generated
and executed.