bootstrap: Ship git-version-gen with the tarball.
bootstrap.sh is already shipped and when rebootstrapping, it will call git-version-gen
internally when processing configure.ac.

In order to know when someone has rebootstrapped, we modify the tarball version to mark it
as rebootstrapped. Not 100% sure this later stage is desired or if we should remove
the .tarball-version file instead.