cli: don't accidentaly set O_NDELAY on stderr
Loading module-cli could have the effect of setting O_NDELAY on stderr,
because it was just a dup'ed fd of stdin which module-cli sets O_NDELAY
for and which flag is shared between all dupes.

Instead of using stdin/stdout directly we now open a new file descriptor
for the controlling terminal, which is equally useful as stdin/stdout
but gives a new file that does not share O_NDELAY with stdin/stdout.

This solves a problem where when running pulseaudio -C resulted in
log output being truncated since stdio does not really handle O_NDELAY
that well in on its fds.