Simplify Makefile.am handling of ALSA-related files.
Instead of declaring extra variables for the ALSA profiles and PATHS, and
using EXTRA_DIST for the udev rule, use the dist_ prefix when declaring
them for conditional installation.

This relies on the fact that automake *is* smart enough to know that there
exist *no* conditional dist, and will always distribute all of them.